Move getPrincipalNameToDisplay logic into service to share
