+++ Fabric.Authorization.AccessControl/src/app/modules/access-control/custom-group/custom-group.component.spec.ts
@@ -353,62 +353,6 @@ describe('CustomGroupComponent', () => {
     });
   });
 
-  describe('getPrincipalNameToDisplay', () => {
-    it('should not append tenant alias if principal is a user', () => {
-      // arrange
-      const principalName = 'azureuser@tenant.com';
-      const principal = {
-        subjectId: principalName,
-        principalType: 'user',
-        tenantId: 'tenantId',
-        tenantAlias: 'tenantAlias',
-        identityProviderUserPrincipalName: principalName,
-      };
-
-      // act
-      const result = component.getPrincipalNameToDisplay(principal);
-
-      // assert
-      expect(result).toBe(principalName);
-    });
-
-    it('should not append tenant alias if principal is a group, and tenant alias is not present', () => {
-      // arrange
-      const principalName = 'azure group';
-      const principal = {
-        subjectId: principalName,
-        principalType: 'group',
-        tenantId: 'tenantId',
-        identityProviderUserPrincipalName: principalName,
-      };
-
-      // act
-      const result = component.getPrincipalNameToDisplay(principal);
-
-      // assert
-      expect(result).toBe(principalName);
-    });
-
-    it('should append tenant alias if principal is a group, and tenant alias is present', () => {
-      // arrange
-      const principalName = 'azure group';
-      const tenantAlias = 'alias';
-      const principal = {
-        subjectId: principalName,
-        principalType: 'group',
-        tenantId: 'tenantId',
-        tenantAlias: tenantAlias,
-        identityProviderUserPrincipalName: principalName,
-      };
-
-      // act
-      const result = component.getPrincipalNameToDisplay(principal);
-
-      // assert
-      expect(result).toBe(principalName + '@' + tenantAlias);
-    });
-  });
-
   describe('getAdGroupNameToDisplay', () => {
     it('should append tenant alias if not null', () => {
       const groupName = 'group-name';
